TimeOnterval as double
because float overflows after just a day :sweat_smile:

+++ OverAnimate.h
@@ -4,7 +4,7 @@
 class AnimationSystem;
 
 /// Time in seconds.
-typedef float TimeInterval;
+typedef double TimeInterval;
 
 /*!
  * @class Animation
